Removed getreferences() from crds.client package altogether in favor of maintaining a single "heavy" version in crds proper.  The spirit of crds.client is now to support jsonrpc level services only.
Added logging to the core jsonrpc service code client.proxy so that every network call can be easily tracked.

Added an observatory parameter to getreferences() which defaults to jwst.   This is needed to generate multi-project file paths, particularly for the location of cached server info.   It is also needed to support the multiple-projects-per-server scenario.   In principle observatory can be computed from the parameters but in practice it is simpler just to specify it directly.  Eliminated config.get_crds_offline_observatory() since it's no longer needed.

Refactored the "heavy" getreferences() for better sharing of common functionality between "local" and "remote" modes.


git-svn-id: https://aeon.stsci.edu/ssb/svn/crds/trunk@484 0d8f46c8-9b30-49d0-b470-0f6283dc92e8