Cleaned up HST parkey expansions / substitutions to remove file clutter *_expansions.dat.
Added HST rmap-generation-time parkey relevance checks so that rmaps are generated with N/A where appropriate rather than clustering on bogus irrelevant randomly set values.   This is intended to simplify rmaps so that automatic "refactoring" and new reference file insertion is coherent.  How to match on irrelevant parameters or N/A during refactoring is confusing;  should a specific value in a new reference match an rmap N/A or not?   Yes,  at least it should not hurt matching even if it complicates the rmap.  Should an N/A in a reference match a specific value in an rmap,  or not?  No, in general,  it should not.   This is why irrelevant values in rmaps need to be converted to N/A as soon as it's known.   This makes it possible to exclude a reference which is not providing any guarantees while obtaining N/A==N/A matches where we're certain the parkey is always irrelevant in the context of the others.

Added CDBS hacks for WFC3 FLSHFILE for FLASHCUR values OFF,ZERO,ZEROCUR not properly represented anywhere;  in these cases FLSHFILE is actually irrelevant.  CDBS still recommends something in these cases though so testing stills shows a "something vs. N/A" mismatch.   Haven't decided how to hide this false positive.





git-svn-id: https://aeon.stsci.edu/ssb/svn/crds/trunk@728 0d8f46c8-9b30-49d0-b470-0f6283dc92e8