Added rmap "extra_keys" to result of get_required_parkeys().  I think I omitted it in the past because all the extra_keys were already in parkey because of the way the HST rmaps were generated.
Basically the extra_key values were not represented in the CDBS matching database used to generate rmaps,  but probably should have been since they're used to assign references.  I added those parameters to generated rmaps as N/A,  thinking perhaps later we'd define real values and make the rmaps what CDBS should have been.  Dumb idea,  doesn't really help,  adds clutter.  Other than that,  doesn't hurt either.

As long as it doesn't cause problems,  the original intent of "extra_keys" was to support parameters which *are not* included in parkey.   This change was motivated by a need to supply missing parameters to an rmap_relevance expression;  so parameters not used for matching per-say,  but for auxiliary functions.  

git-svn-id: https://aeon.stsci.edu/ssb/svn/crds/trunk@1512 0d8f46c8-9b30-49d0-b470-0f6283dc92e8