Changes to infrastructure so crds.bestrefs can run on a test .rmap instead of a cached .pmap.
Modified core bestrefs logic in rmap.py so that ReferenceMapping implements a degenerate get_best_references() interface and now consolidates the error handling functionality instead of InstrumentMapping.  Updated unit tests in tobs broken by this change in exception handling.

Modified several routines in client.api.py to pass the basename of the context to the server.

Modified CrdsError in client.proxy.py to support multiple parameters which are space-joined akin to print().

git-svn-id: https://aeon.stsci.edu/ssb/svn/crds/trunk@2038 0d8f46c8-9b30-49d0-b470-0f6283dc92e8