Enhanced newcontext script and functions to support a pseudo-rmap notation used to set values to N/A.
Normally crds.newcontext takes an existing pmap and a list of new
rmaps to inject in order to create a new pmap and related imaps.  With
this enhancement, one can also specify e.g. miri_dflat_n/a and
newcontext will set the MIRI DFLAT value to N/A instead of an
rmap.