Significant overhaul of certifier to factor generic .tpn functions out of both the validators module and observatory packages.
One significant change was the elimination of "conditional .tpn files"
used for HST's STIS FLTFILE and PFLTFILE types by exploiting the
newer "conditionally required" constraint feature.   This enables those
types to conform to the common model of .tpn files corresponding directly
to rmaps and reference file types.   This simplifies .tpn lookups and
also makes them independent of the actual reference files being certified.
Under the new approach,  the constraints that are applicable to a reference
file may vary with conditional logic,  but the files that define the constaints
have a simple uniform layout based on instrument and type.