Added support to crds.io.fits to refer to FITS array extensions by number, e.g. get_array_properities() can now specify data for HDU #1 as:
     "EXT1"
     "EXTENSION1"
     "1"

This paves the way for .tpn checks written using the above names in addition to
normal HDU names, referred to in constraint expressions normally as
e.g. "EXT1_ARRAY".

The preferred form of checks should always be by HDU name not number, but it's
not uncommon for extension names to be multiply defined or missing.