Switched SYNPHOT specs / baseline rmaps from rmap_relevance which conditionally forces bestref values to N/A to rmap_omit which drops the keyword entirely with the intent of preventing SYNPHOT keywords from being added to CAL bestrefs dataset headers.
Added the comment_parkeys rmap header field which enables simple declaration of
parkeys which are added to match tuples but always ignored during bestrefs.
This enables the SYNPHOT TMCTAB and TMTTAB rmaps to encode most of the
information in the TMC and TMT FITS references and display it on the web.
This is different than classic parkey_relevance because (a) it's cleaner to
declare and (b) the value of the comment keywords is placed in the rmap
instead of being normalized to N/A.

Dropped prototype rmap update_raw_parkeys field.